done backend

+++ app/Http/Controllers/SiswaController.php
@@ -15,14 +15,30 @@ class SiswaController extends Controller
     public function index()
     {
         $siswa = DB::table('tb_siswa')
-                ->join('tb_kelas', 'tb_kelas.id_kelas', '=', 'tb_siswa.id_kelas')
+                ->leftJoin('tb_kelas', 'tb_kelas.id_kelas', '=', 'tb_siswa.id_kelas')
                 ->select('tb_siswa.*', 'tb_kelas.nama_kelas', 'tb_kelas.grup_kelas')
                 ->get();
-
         return view(
             'page/siswa/index',
             [
                 'siswa' => $siswa
+            ]
+        );
+    }
+    
+    public function verified()
+    {
+        $kelas = Kelas_Model::get();
+        $siswa = DB::table('tb_siswa')
+                ->leftJoin('tb_kelas', 'tb_kelas.id_kelas', '=', 'tb_siswa.id_kelas')
+                ->select('tb_siswa.*', 'tb_kelas.nama_kelas', 'tb_kelas.grup_kelas')
+                ->where('tb_siswa.status_daftar', '1')
+                ->get();
+        return view(
+            'page/siswa/verified_student',
+            [
+                'siswa' => $siswa,
+                'kelas' => $kelas,
             ]
         );
     }
@@ -31,7 +47,7 @@ class SiswaController extends Controller
         $kelas = Kelas_Model::all();
         return view(
             'page/siswa/form',
-            [   
+            [
                 'url' => 'siswa.store',
                 'kelas' => $kelas
             ]
@@ -41,6 +57,7 @@ class SiswaController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'nama_siswa'         => 'required',
+            'nisn'         => 'required|numeric|unique:tb_siswa,nis',
             'gender_siswa'         => 'required',
             'nohp_siswa'         => 'required',
             'tempat_lahir_siswa'         => 'required',
@@ -56,11 +73,24 @@ class SiswaController extends Controller
                 ->withErrors($validator)
                 ->withInput();
         } else {
-            $foto = $request->file('foto_siswa'); 
-            $filename = time() . "." . $foto->getClientOriginalExtension(); 
-            $foto->move('backend/img/siswa/', $filename);
+            if ($request->hasFile('foto_siswa')) {
+                $foto = $request->file('foto_siswa'); 
+                $filename = time() . "." . $foto->getClientOriginalExtension(); 
+                $foto->move('backend/img/siswa/', $filename);
+            }else{
+                $filename = "siswa.png";
+            }
+            $thn = $request->input('nis');
+            $row = DB::table('tb_siswa')
+                    ->where('ta', $thn)
+                    ->count();
+            $order = sprintf("%03s", $row+1);
+            $nis = strval($thn).strval($order);
 
             $siswa->nama_siswa = $request->input('nama_siswa');
+            $siswa->ta = $request->input('nis');
+            $siswa->nis = $nis;
+            $siswa->nisn = $request->input('nisn');
             $siswa->gender_siswa = $request->input('gender_siswa');
             $siswa->nohp_siswa = $request->input('nohp_siswa');
             $siswa->tempat_lahir_siswa = $request->input('tempat_lahir_siswa');
@@ -94,6 +124,7 @@ class SiswaController extends Controller
     {
         $validator = Validator::make($request->all(),[
             'nama_siswa'         => 'required',
+            'nisn'         => 'required|numeric|unique:tb_siswa,nis',
             'gender_siswa'         => 'required',
             'nohp_siswa'         => 'required',
             'tempat_lahir_siswa'         => 'required',
@@ -111,14 +142,19 @@ class SiswaController extends Controller
         }else{
             // cek apakah user kirim gambar lagi/tidak 
             if ($request->hasFile('foto_siswa')) {
-                // cari nama foto lama lalu hapus 
-                unlink('backend/img/siswa/' . $siswa->foto_siswa); 
+                // cek dulu apakah gambar bukan gambar defaut
+                if($siswa->foto_siswa != 'siswa.png'){
+                    // cari nama foto lama lalu hapus 
+                    unlink('backend/img/siswa/' . $siswa->foto_siswa); 
+                }
                 $foto = $request->file('foto_siswa'); 
                 $filename = time() . "." . $foto->getClientOriginalExtension(); 
                 $foto->move('backend/img/siswa/', $filename); 
                 $siswa->foto_siswa = $filename; 
-            }
+            }            
             $siswa->nama_siswa = $request->input('nama_siswa');
+            $siswa->nis = $request->input('nis');
+            $siswa->nisn = $request->input('nisn');
             $siswa->gender_siswa = $request->input('gender_siswa');
             $siswa->nohp_siswa = $request->input('nohp_siswa');
             $siswa->tempat_lahir_siswa = $request->input('tempat_lahir_siswa');
@@ -136,8 +172,11 @@ class SiswaController extends Controller
 
     public function destroy(Siswa_Model $siswa)
     {
-        $siswa_file = $siswa->foto_siswa; 
-        unlink('backend/img/siswa/' . $siswa_file); 
+        // cek dulu apakah gambar bukan gambar defaut
+        if($siswa->foto_siswa != 'siswa.png'){
+            // cari nama foto lama lalu hapus 
+            unlink('backend/img/siswa/' . $siswa->foto_siswa); 
+        }
         $siswa->forceDelete();
         
         return redirect()
@@ -145,6 +184,80 @@ class SiswaController extends Controller
             ->with('message', 'Data berhasil dihapus');
     }
 
+    public function terdaftar(Request $request)
+    {
+        // dd($request->all());=
 
-    
+        DB::table('tb_user')
+            ->insert([
+                'username' => $request->nama_siswa,
+                'password' => Hash::make(12345),
+                'id_s' => $request->id_siswa,
+                'level' => '6'
+            ]);
+
+        $siswa = DB::table('tb_siswa')
+            ->where('id_siswa', $request->id_siswa)
+            ->update(['status_daftar' => 1]);
+            
+        
+        return json_encode($siswa);
+
+
+    }
+
+    public function tdk_terdaftar(Request $request)
+    {
+        DB::table('tb_user')
+            ->where('id_s', '=', $request->id_siswa) 
+            ->where('level', '=', 6) 
+            ->delete();
+
+        $siswa = DB::table('tb_siswa')
+            ->where('id_siswa', $request->id_siswa)
+            ->update(['status_daftar' => 0]);
+
+
+        return json_encode($siswa);
+    }
+
+    public function exportExcel(Request $request)
+    {
+        $kelas = DB::table('tb_kelas')->get();
+        if ($request->kelas == "0") {
+            $data['siswa'] =  DB::table('tb_siswa')
+                        ->join('tb_kelas', 'tb_siswa.id_kelas', '=', 'tb_kelas.id_kelas')
+                        ->get();
+
+            $data['rekap'] = 'semua';
+
+            return view('page/siswa/cetak', $data);
+        }
+        elseif ($request->kelas == "jumlah") {
+            $data['siswa'] = DB::table('tb_siswa')
+                        ->join("tb_kelas", "tb_siswa.id_kelas", "=", "tb_kelas.id_kelas")
+                        ->select(
+                            'tb_siswa.id_kelas', 
+                            'tb_kelas.*',
+                            DB::raw('COUNT(CASE WHEN gender_siswa="Pria" THEN 1  END) As Male'),
+                            DB::raw('COUNT(CASE WHEN gender_siswa="Wanita" THEN 1  END) As Female'),
+                            )
+                        ->where('tb_siswa.status_daftar',  '1')
+                        ->groupBy('tb_siswa.id_kelas')
+                        ->get();
+
+            $data['rekap'] = 'jml'; 
+            return view('page/siswa/cetak', $data);
+        }
+        else{
+            $data['siswa'] =  DB::table('tb_siswa')
+                        ->join('tb_kelas', 'tb_siswa.id_kelas', '=', 'tb_kelas.id_kelas')
+                        ->where("tb_siswa.status_daftar", "=", '1')
+                        ->where("tb_siswa.id_kelas", "=", $request->kelas)
+                        ->get();
+            // dd($siswa);
+            $data['rekap'] = ''; 
+            return view('page/siswa/cetak', $data);
+        }
+    }
 }
